Move webserver and its proxy apps to host network
This was necessary so that connecting clients appear with their IP rather than the docker network's IP. If we didn't do this, everyone would appear to have the sdame IP which would make IP filtering and ddos mitigation completely ineffective. This does reduce security as there is no more a network isolation layer for each web app, but then again, this was necessary.