Add a backups user
The backups user is basically just admin, except in the sshd_config we removed ForceCommand for it so that it could use rsync without issues. It also just made more sense to isolate the admin which a real person would use and a backups account that is likely controlled by automated software.