Fallbacks for LXCFS mounts and optional
It turns out LXCFS's slabinfo does not exist for all machines, which prevented the project from running. Instead of getting rid of it, I decided to turn it into variables with a fallback value. This should prevent a similar issue from occurring in the future. Additionally LXCFS has now been made optional. It can be disabled inside .env.