Comment out discord status bot
My reasoning is that you should NOT report the status of services on the same machine you are running the services on because if the machine goes offline, the bot won't be reporting anything which defeats its whole purpose. It has been disabled rather than removed, so those who prefer to run it as part of the docker cluster can still do so with a single line reduction.