Remove syslog and fail2ban
There is a race condition with containers that define syslog. It's been reported several times; there's an open issue since 2016, and yet, the Docker devs have not addressed it and probably never will. Since syslog is forever broken and fail2ban needs log files, both of them are gone now.